LayoutTests/http/tests/security/contentSecurityPolicy/1.1/base-uri-deny-expected.txt

CONSOLE MESSAGE: Refused to change the document base URL to http://127.0.0.1:8000/security/contentSecurityPolicy/1.1/resources/base-href/ because it does not appear in the base-uri directive of the Content Security Policy.
ALERT: This is a safe script.
